Added Conditions for Non-Empty Info, Showtime Reservation Link, and Driver Update
Enhanced the scraper functionality by implementing conditions to verify that the information to be retrieved is not empty. This ensures that only valid and non-empty data is collected during the scraping process, improving the accuracy and reliability of the results.

Additionally, the scraper now includes the capability to fetch the reservation link for each showtime. This allows users to directly access the booking page for a specific showtime, enhancing the user experience and facilitating the reservation process.

Furthermore, the Chrome driver has been updated to the latest version. This update ensures compatibility with the latest features and improvements, providing a more efficient and reliable scraping experience.